+++ apps/web/src/app-pages/mypage/settings/policy/ui/PublicPrivacyPolicyPage.tsx
@@ -14,7 +14,7 @@ const PublicPrivacyPolicyPage = () => {
           hasLeftIcon: true,
         }}
       />
-      <PolicyDetailItem policyId="PrivacyPolicy" />
+      <PolicyDetailItem policyId="PersonalInfoPolicy" />
     </div>
   );
 };
