+++ apps/web/src/app/api/proxy/[...path]/route.ts
@@ -189,7 +189,7 @@ function parseSetCookie(cookie: string) {
 }
 
 function toProxyCookiePath(name: string, path?: string) {
-  if (name === 'oauth_state') return '/';
+  if (name === 'oauth_state') return '/api/auth';
   if (!path || path === '/') return path ?? '/';
   if (path.startsWith('/api/proxy')) return path;
   return `/api/proxy${path}`;
